Aplicacion/Avance 3-Monitor-recursos/requirements.txt
ttkbootstrap>=1.10.1
psutil>=5.9.8
Pillow>=10.0.0
